Trait embedded_can::blocking::Can

pub trait Can {
    type Frame: Frame;
    type Error;
    fn try_write(&mut self, frame: &Self::Frame) -> Result<(), Self::Error>;

    fn try_read(&mut self) -> Result<Self::Frame, Self::Error>;
}

A blocking CAN interface that is able to transmit and receive frames.

Associated Types

type Frame: Frame

Associated frame type.

type Error

Associated error type.

Loading content...
Required methods

fn try_write(&mut self, frame: &Self::Frame) -> Result<(), Self::Error>

Puts a frame in the transmit buffer. Blocks until space is available in the transmit buffer.

fn try_read(&mut self) -> Result<Self::Frame, Self::Error>

Blocks until a frame was received or an error occured.

Loading content...