embedded_can
Struct ExtendedId

pub struct ExtendedId(/* private fields */);

Extended 29-bit CAN Identifier (0..=1FFF_FFFF).

Implementations

impl ExtendedId

pub const ZERO: Self

CAN ID 0, the highest priority.

pub const MAX: Self

CAN ID 0x1FFFFFFF, the lowest priority.

pub const fn new(raw: u32) -> Option<Self>

Tries to create a ExtendedId from a raw 32-bit integer.

This will return None if raw is out of range of an 29-bit integer (> 0x1FFF_FFFF).

pub const unsafe fn new_unchecked(raw: u32) -> Self

Creates a new ExtendedId without checking if it is inside the valid range.

Safety

Using this method can create an invalid ID and is thus marked as unsafe.

pub fn as_raw(&self) -> u32

Returns this CAN Identifier as a raw 32-bit integer.

pub fn standard_id(&self) -> StandardId

Returns the Base ID part of this extended identifier.