Trait embedded_can::Frame

pub trait Frame: Sized {
    fn new(id: impl Into<Id>, data: &[u8]) -> Result<Self, ()>;

    fn new_remote(id: impl Into<Id>, dlc: usize) -> Result<Self, ()>;

    fn is_extended(&self) -> bool;

    fn is_remote_frame(&self) -> bool;

    fn id(&self) -> Id;

    fn dlc(&self) -> usize;

    fn data(&self) -> &[u8];

    fn is_standard(&self) -> bool { ... }

    fn is_data_frame(&self) -> bool { ... }
}

A CAN2.0 Frame

Required methods

fn new(id: impl Into<Id>, data: &[u8]) -> Result<Self, ()>

Creates a new frame. Returns an error when the data slice is too long.

fn new_remote(id: impl Into<Id>, dlc: usize) -> Result<Self, ()>

Creates a new remote frame (RTR bit set). Returns an error when the data length code (DLC) is not valid.

fn is_extended(&self) -> bool

Returns true if this frame is a extended frame.

fn is_remote_frame(&self) -> bool

Returns true if this frame is a remote frame.

fn id(&self) -> Id

Returns the frame identifier.

fn dlc(&self) -> usize

Returns the data length code (DLC) which is in the range 0..8.

For data frames the DLC value always matches the length of the data. Remote frames do not carry any data, yet the DLC can be greater than 0.

fn data(&self) -> &[u8]

Returns the frame data (0..8 bytes in length).

Loading content...
Provided methods

fn is_standard(&self) -> bool

Returns true if this frame is a standard frame.

fn is_data_frame(&self) -> bool

Returns true if this frame is a data frame.

Loading content...