embedded_can
Trait Error

pub trait Error: Debug {
    // Required method
    fn kind(&self) -> ErrorKind;
}

CAN error

Required Methods

fn kind(&self) -> ErrorKind

Convert error to a generic CAN error kind

By using this method, CAN errors freely defined by HAL implementations can be converted to a set of generic serial errors upon which generic code can act.